requirements.txt
aiogram
google-generativeai